==> notebooks/1-clean_data.ipynb <==
# Clean Data + Reorder Data

- Actually we have the data in different languages so we have to create a program to translate the different data

import pandas as pd
# We will try to use a translator that aleady exist
from googletrans import Translator, constants

translator = Translator()
# Await each time that translator function is done from spanish to english

data = pd.read_excel("../data/1-Raw/Forms-Results-Raw.xlsx")

data.head()

async def TranslateWithLibrary(column):
    idx = 0
    for occup in data[column]:
        translation = await(translator.translate(occup, src="es", dest="en"))
        data.loc[idx,column] = translation.text
        idx += 1

def TranslateYesNo(column):
    idx = 0
    for productive in data[column]:
        output = ''
        if productive == 'Si' or productive == 'Yes':
            output = 'Yes'
        else:
            output = 'No'

        data.loc[idx,column] = output
        idx += 1

# Translate Occupation
await(TranslateWithLibrary('Occupation'))

# Translate Country 
await(TranslateWithLibrary('Country'))

# Translate Do you feel more productive when using AI tools?
TranslateYesNo('Do you feel more productive when using AI tools?')

# Translate Do you believe overuse of AI can negatively affect mental health or problem-solving skills?
TranslateYesNo('Do you believe overuse of AI can negatively affect mental health or problem-solving skills?')

# Translate Do you know someone that use AI as conversational partner?
TranslateYesNo('Do you know someone that use AI as  conversational partner?')

- Null data is only in Application question so its okay to not modify it
- We have to modify the data in a question do that its not a list and we need it as a list

# Column What do you primarily use AI tools for? create as a tuple
idx = 0
for usages in data['What do you primarily use AI tools for?']:
    new_list = usages.split(',')
    data['What do you primarily use AI tools for?'][idx] = []
    for usage in new_list:
        data.loc[idx, 'What do you primarily use AI tools for?'].append(usage)
    idx += 1

data.keys()

# Export translated data

data.to_csv('../data/2-Processed/Forms-Results-Translated.csv', index=False)
