==> notebooks/3-compute_risk_data.ipynb <==
# Compute Risk

- Set the necessary values for each column to pass from a string to a number
- Compute the total risk and set the limit value using the using Wellbeing scale and the likert notation (read final report for more detail)

import pandas as pd
import copy

data = pd.read_csv("../data/2-Processed/Forms-Results-Translated.csv")
data.tail()

data.keys()

- Erase unused columns for risk

data = data.drop(columns=['Age',
                          'Occupation',
                          'Gender',
                          'Country',
                          'What application with AI would you like to exist?'])

- Divide the questions usyng Ryff's scale and we will compute from that scale

ryffmap_output = {
    'Self-Acceptance' : {
        'Relying on AI helps reduce my stress.' : 0,
        'AI tools make me second-guess my own knowledge or decisions.' : 0,
        'Using AI makes me feel mentally lazy.' : 0
    },
    'Positive Relations with Others' : {
        'Do you know someone that use AI as  conversational partner?' : 0,
        'Its recomendable for people to use AI as conversational partner' : 0
    },
    'Autonomy' : {
        'I feel more confident in solving problems on my own without AI assistance.' : 0,
        'Do you believe overuse of AI can negatively affect mental health or problem-solving skills?' : 0,
        'AI tools make me second-guess my own knowledge or decisions.' : 0
    },
    'Environmental Mastery' : {
        'Do you feel more productive when using AI tools?' : 0,
        'How often do you use AI tools?' : 0,
        'Relying on AI helps reduce my stress.' : 0
    },
    'Purpose in Life' : {
        'What do you primarily use AI tools for?' : 0
    },
    'Personal Growth' : {
        'Do you believe overuse of AI can negatively affect mental health or problem-solving skills?' : 0,
        'Using AI makes me feel mentally lazy.' : 0,
        'What do you primarily use AI tools for?' : 0
    }
}

list_ryffmap_output = []
list_ryffmap_output = [copy.deepcopy(ryffmap_output) for _ in data['ID']]

- Use the likert notation to compute the scale 

likert_map = {
    # Strongly agree to strongly disagree
    'Strongly Agree': 5,
    'Agree': 4,
    'Neutral': 3,
    'Disagree': 2,
    'Strongly Disagree': 1,

    # Yes No question
    'Yes': 4,
    'No': 2,

    # Using rate of the AI
    'Daily': 5,
    'Several times a week': 4,
    'Occasionally': 3,
    'Rarely': 2,
    'Never': 1,
}

- We will compute by dimensions (each column of the question its a dimension)

def ComputeRiskValue(column):
    # Check column and get the likert value
    idx = 0
    for element in data[column]:
        # Get the map value
        value = likert_map[element]
        
        # Check on the list of outputs take one output and compute the data
        for dimension, subelement in list_ryffmap_output[idx].items():
            if column in subelement:
                list_ryffmap_output[idx][dimension][column] = value
           
        idx += 1
    

ComputeRiskValue('How often do you use AI tools?')
ComputeRiskValue('Do you feel more productive when using AI tools?')
ComputeRiskValue('Relying on AI helps reduce my stress.')
ComputeRiskValue('AI tools make me second-guess my own knowledge or decisions.')
ComputeRiskValue('I feel more confident in solving problems on my own without AI assistance.')
ComputeRiskValue('Using AI makes me feel mentally lazy.')
ComputeRiskValue('Do you believe overuse of AI can negatively affect mental health or problem-solving skills?')
ComputeRiskValue('Do you know someone that use AI as  conversational partner?')
ComputeRiskValue('Its recomendable for people to use AI as conversational partner')

print(list_ryffmap_output[0])
print(list_ryffmap_output[1])

- Now that we have the computations of each dimension we will do the final computation to get the score of each dimension

ryffmap_rate = {
    'Self-Acceptance' :  0,
    'Positive Relations with Others': 0,
    'Autonomy' :  0,
    'Environmental Mastery' : 0,
    'Purpose in Life' : 0,
    'Personal Growth' : 0
}

list_rate_output = []    
list_rate_output = [copy.deepcopy(ryffmap_rate) for _ in data['ID']]

# Loop through all the list of rates
for idx, output in enumerate(list_ryffmap_output):
    # Check all the elements 
    for dimension, subelement in output.items():
        count = 0
        result = 0
        for element in subelement:
            result += (list_ryffmap_output[idx][dimension][element])
            count += 1
        
        list_rate_output[idx][dimension] =  result/count

print(list_rate_output[0])
print(list_rate_output[1])

- Transform it to a new data

risk_rate = pd.DataFrame(list_rate_output)

risk_rate['ID'] = data['ID']

risk_rate.head()

- Plot

import plotly.express as px
plotRate = risk_rate

plotRate = plotRate.drop(columns =['ID', 'Purpose in Life'])


fig = px.imshow(plotRate, 
                 labels=dict(x="Ryff Scale Questions", y="ID", color="Risk"),
                text_auto=True,  zmin = 0, zmax= 4,height=1150)
fig.update_xaxes(side="top")
fig.show()

- Detect Risk

def DetectRisk(row):
    row_selected = plotRate.iloc[row]
    result = 0
    for element in row_selected:
        result+=element

    result /=4
    print(f"Current Risk: {result}")
    if(result < 2.5):
        print("Recommended psychologist")
    else:
        print("Not recommended psychologist")


DetectRisk(50)

# Export to another file

- Export to a basic file that will be used to create the dashboard basic documentation

data.to_csv('../data/2-Processed/Forms-Results-Compute_Risk.csv', index=False)  
plotRate.to_csv('../data/2-Processed/Computed_Risk_Values.csv', index=False)
