==> notebooks/4-studies_and_hypothesis.ipynb <==
import pandas as pd
from scipy.stats import ttest_ind, f_oneway
import plotly.express as px

Read documents

data = pd.read_csv("../data/2-Processed/Forms-Results-Translated.csv")
data.sample(3)


risk = pd.read_csv("../data/2-Processed/Computed_Risk_Values.csv")
risk.head()

# Studies

- Different use of compute risk by gender

# Put the gender column from one to other

genderRelated = risk
genderRelated['Gender'] = data['Gender']

genderRelated.head()

newRelated = genderRelated.groupby('Gender').mean(numeric_only=True)

fig = px.imshow(newRelated, 
                 labels=dict(x="Ryff Scale Questions", y="Gender", color="Risk"),
                text_auto=True,  zmin = 0, zmax= 4,height=500)
fig.update_xaxes(side="top")
fig.show()

- Different use of compute risk by Age

ageRelated = risk
ageRelated['Age'] = data['Age']

ageRelated.head()

newRelated = ageRelated.groupby('Age').mean(numeric_only=True)

fig = px.imshow(newRelated, 
                 labels=dict(x="Ryff Scale Questions", y="Age", color="Risk"),
                text_auto=True,  zmin = 0, zmax= 4,height=900)
fig.update_xaxes(side="top")
fig.show()

# Hypothesis

- Frequent users of AI tools are more likely to report beeing less stress.

Null Hypothesis (H₀): There is no difference in self-reported less stress between frequent and infrequent users.

Alternative Hypothesis (H₁): Frequent users report significantly higher feelings of less stress.

likert_map = {
    'Strongly Disagree': 1,
    'Disagree': 2,
    'Neutral': 3,
    'Agree': 4,
    'Strongly Agree': 5
}

data['stress_reduction_score'] = data['Relying on AI helps reduce my stress.'].map(likert_map)
data['usage_group'] = data['How often do you use AI tools?'].map({
    'Daily': 'Frequent',
    'Several times a week': 'Frequent',
    'Occasionally': 'Infrequent',
    'Rarely': 'Infrequent',
    'Never': 'Infrequent'
})


df_clean = data.dropna(subset=['stress_reduction_score', 'usage_group'])

frequent = df_clean[df_clean['usage_group'] == 'Frequent']['stress_reduction_score']
infrequent = df_clean[df_clean['usage_group'] == 'Infrequent']['stress_reduction_score']

t_stat, p_value = ttest_ind(frequent, infrequent, equal_var=False)

p_one_tailed = p_value / 2 if t_stat > 0 else 1 - (p_value / 2)

confidence_level = (1 - p_one_tailed) * 100

print(f"T-statistic: {t_stat:.4f}")
print(f"One-tailed p-value: {p_one_tailed:.4f}")
print(f"Confidence Level: {confidence_level:.2f}%")

if p_one_tailed < 0.05:
    print("✅ Statistically significant: Frequent users have less stress.")
else:
    print("❌ Not statistically significant: No strong evidence that frequent users have less stress.")
